nmr|tests: Relocate tests to sibling directory of tested files
