+++ packages/nmr/src/__tests__/help.test.ts
@@ -4,7 +4,7 @@ import path from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
-import { generateHelp } from '../src/help.js';
+import { generateHelp } from '../help.ts';
 
 describe(generateHelp, () => {
   it('includes usage line', () => {
