+++ packages/nmr/src/helpers/__tests__/hook-name.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { isHookName } from '../../src/helpers/hook-name.js';
+import { isHookName } from '../hook-name.ts';
 
 describe(isHookName, () => {
   it.each(['build:pre', 'build:post', 'test:pre', 'test:post', 'some:nested:post', 'x:pre', 'x:post'])(
