+++ packages/nmr/vitest.config.ts
@@ -1,0 +1,1 @@
+export { default } from '../../config/vitest.config.ts';
